// src/Interfaces/enums/EImageType.ts
export enum EImageType {
    STILL = 'STILL',
    SHOOTING = 'SHOOTING',
    POSTER = 'POSTER',
    FAN_ART = 'FAN_ART',
    PROMO = 'PROMO',
    CONCEPT = 'CONCEPT',
    WALLPAPER = 'WALLPAPER',
    COVER = 'COVER',
    SCREENSHOT = 'SCREENSHOT',

}
export enum EImageTypeRu {
    STILL = 'Кадры',
    SHOOTING = 'Изображения со съемок',
    POSTER = 'Постеры',
    FAN_ART = 'Фан-арты',
    PROMO = 'Промо',
    CONCEPT = 'Концепт-арты',
    WALLPAPER = 'Обои',
    COVER = 'Обложки',
    SCREENSHOT = 'Скриншоты',
}